+++ basics/src/main/kotlin/de/thermondo/basics/Basics2.kt
@@ -5,6 +5,6 @@ package de.thermondo.basics
  * Functions by default are void.
  * Function naming is done in lower camel case.
  *
- * Let's define a function called hello world that prints "Hello World! to the console"
- * println() can be used to print messages to the consol
+ * Let's define a function called hello world that prints "Hello World!" to the console
+ * println() can be used to print messages to the console
  */
