+++ settings.gradle.kts
@@ -1,3 +1,4 @@
 rootProject.name = "Let's Learn Kotlin"
 include("basics")
 include("solutions")
+include("test-utils")
