+++ solutions/src/main/kotlin/de/thermondo/solutions/basics/Basics2Solution.kt
@@ -5,8 +5,8 @@ package de.thermondo.solutions.basics
  * Functions by default are void.
  * Function naming is done in lower camel case.
  *
- * Let's define a function called hello world that prints "Hello World! to the console"
- * println() can be used to print messages to the consol
+ * Let's define a function called hello world that prints "Hello World!" to the console
+ * println() can be used to print messages to the console
  */
 fun helloWorld() {
     println("Hello World!")
