--- solutions/src/test/kotlin/de/thermondo/solutions/basics/Basic3SolutionTest.kt
@@ -1,5 +1,0 @@
-package de.thermondo.solutions.basics
-
-import org.junit.jupiter.api.Assertions.*
-
-internal class Basic3SolutionTest
