+++ solutions/src/test/kotlin/de/thermondo/solutions/basics/Basics2SolutionTest.kt
@@ -1,9 +1,9 @@
-package de.thermondo.basics
+package de.thermondo.solutions.basics
 
 import de.thermondo.test.utils.BasePrintlnTest
 import kotlin.test.Test
 
-class Basics2Test : BasePrintlnTest() {
+class Basics2SolutionTest : BasePrintlnTest() {
     @Test
     fun `Validate helloWorld function`() {
         helloWorld()
